move & refactor routes, controller to namespace
