+++ app/views/products/heading/_your_products.html.erb
@@ -7,7 +7,7 @@
   </div>
   <div class="govuk-grid-column-one-third">
     <div class="opss-text-align-right govuk-!-margin-bottom-8">
-      <%= link_to "Create a product record", new_product_duplicate_check_path, class: "govuk-link govuk-link--no-visited-state govuk-!-font-size-19 opss-text-underline-offset" %>
+      <%= link_to "Create a product record", duplicate_check_products_path, class: "govuk-link govuk-link--no-visited-state govuk-!-font-size-19 opss-text-underline-offset" %>
     </div>
   </div>
 </div>
